+++ net/hlssink3/src/playlist.rs
@@ -93,7 +93,7 @@ impl Playlist {
         }
 
         self.playlist_index += 1;
-        self.inner.media_sequence = self.playlist_index as u64 - self.inner.segments.len() as u64;
+        self.inner.media_sequence = self.playlist_index - self.inner.segments.len() as u64;
     }
 
     /// Sets the playlist to started state.
